% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_comprehension.R
\name{list_comprehension}
\alias{list_comprehension}
\title{Pacote que implementa list comprehension utilizando for loop}
\usage{
list_comprehension(expr)
}
\arguments{
\item{expr}{expressão de for loop}
}
\description{
Implementa list comprehension apenas para casos simples utilizando for loop.
}
\examples{

# cria uma lista com o quadrado dos valores ímpares
list_comprehension(for(i in 1:10) if(i \%\% 2 != 0) i^2)

# faz um split e toma tudo aquilo que aparece antes da chave
list_comprehension(for(i in c("bata}ta", "ceno}ura", "laran}ja")) strsplit(i, "}")[[1]][[1]])
}
\author{
Gabriel Pereira
}
